Add support of VendorEvent
The VendorEvent provide a generic and easy way to create
a perf event when existing event types are not appropriated.
Here, we are going to use it to create a PerfEvent that show
the value of any event, scaled to the second.
This allow to see any event in tools using Perf and PerfEvent.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>